Version 1.0 completed

+++ xeon_version/include/program_options.h
@@ -68,7 +68,7 @@ public:
 	 * @param default_value		value to return if the option value is <code>NULL</code>.
 	 * @return					the option integer value or the <em>default_value</em>.
 	 */
-	size_t get_number( const std::string& option, size_t default_value ) const;
+	long get_number( const std::string& option, long default_value ) const;
 
 	/**
 	 * Retrieve the integer value of one of the two options we want to retrieve.
@@ -76,7 +76,7 @@ public:
 	 * @param default_value		value to return if the option value is <code>NULL</code>.
 	 * @return					the option integer value or the <em>default_value</em>.
 	 */
-	size_t get_number( const std::string& option1, const std::string& option2, size_t default_value ) const;
+	long get_number( const std::string& option1, const std::string& option2, long default_value ) const;
 
 private:
 	int argc;
